embeddings: add support for prefixes in embeddings
Some of the embedding models (e.g  nomic-embed-text) are trained to embed for
several purposes, this on the flip side, means that if not given the purpose
for the embedding the retrieval results might and probably will be suboptimal.
This change aims to deal with it by adding a configuration option for chunk and
query prefixes with the option to support more embedding purposes in the future
(i.e clustering).

Some refferences to justify this change:
1. https://huggingface.co/nomic-ai/nomic-embed-text-v1.5#task-instruction-prefixes
   The nomic-embed-text prefixes section
2. https://www.youtube.com/watch?v=76EIC_RaDNw (Don’t Embed Wrong! by Matt
   Williams)

Signed-off-by: Eliran Sinvani <eliransin@gmail.com>